Move the type augmentation to directory-handle-mount.ts
