+++ packages/php-wasm/web/src/lib/directory-handle-mount.ts
@@ -8,6 +8,30 @@ import {
 import { Semaphore, joinPaths } from '@php-wasm/util';
 import { logger } from '@php-wasm/logger';
 import { FilesystemOperation, journalFSEvents } from '@php-wasm/fs-journal';
+import '../types';
+
+declare global {
+	interface FileSystemDirectoryHandle {
+		values: () => AsyncIterable<
+			FileSystemDirectoryHandle | FileSystemFileHandle
+		>;
+	}
+	interface FileSystemFileHandle {
+		move(target: FileSystemDirectoryHandle): Promise<void>;
+		move(name: string): Promise<void>;
+		move(target: FileSystemDirectoryHandle, name: string): Promise<void>;
+		createWritable(): Promise<FileSystemWritableFileStream>;
+	}
+	interface FileSystemWritableFileStream {
+		write(
+			buffer: BufferSource,
+			options?: FileSystemReadWriteOptions
+		): Promise<number>;
+		close(): Promise<void>;
+		seek(offset: number): Promise<void>;
+		truncate(newSize: number): Promise<void>;
+	}
+}
 
 export interface MountOptions {
 	initialSync: {
